HDDS-15800. Move mocks out of TestUpgradeFinalizerActions
